+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/activity/AgreementActivity.kt
@@ -3,10 +3,10 @@ package com.docusign.sdksamplekotlin.activity
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.docusign.sdksamplekotlin.R
-import com.docusign.sdksamplekotlin.fragment.PortfolioFragment
-import com.docusign.sdksamplekotlin.fragment.ClientInvestmentFragment
 import com.docusign.sdksamplekotlin.fragment.AgreementHomeFragment
 import com.docusign.sdksamplekotlin.fragment.AgreementTemplatesFragment
+import com.docusign.sdksamplekotlin.fragment.ClientInvestmentFragment
+import com.docusign.sdksamplekotlin.fragment.PortfolioFragment
 import com.docusign.sdksamplekotlin.model.Client
 import com.google.gson.Gson
 
@@ -18,6 +18,7 @@ class AgreementActivity : AppCompatActivity(), PortfolioFragment.PortFolioFragme
     companion object {
         private val TAG = AgreementActivity::class.java.simpleName
         const val CLIENT_DETAILS = "ClientDetails"
+        const val OPEN_CLIENT_INVESTMENT = "Open client investment"
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -27,36 +28,57 @@ class AgreementActivity : AppCompatActivity(), PortfolioFragment.PortFolioFragme
             client = Gson().fromJson(clientJson, Client::class.java)
         }
         setContentView(R.layout.activity_agreement)
-        var agreementHomeFragment = supportFragmentManager.findFragmentByTag(AgreementHomeFragment.TAG)
+        var agreementHomeFragment =
+            supportFragmentManager.findFragmentByTag(AgreementHomeFragment.TAG)
         if (agreementHomeFragment == null) {
             val fragmentTransaction = supportFragmentManager.beginTransaction()
             agreementHomeFragment = AgreementHomeFragment.newInstance(client)
-            fragmentTransaction.add(R.id.agreement_container, agreementHomeFragment, AgreementHomeFragment.TAG)
+            fragmentTransaction.add(
+                R.id.agreement_container,
+                agreementHomeFragment,
+                AgreementHomeFragment.TAG
+            )
             fragmentTransaction.addToBackStack(AgreementHomeFragment.TAG)
             fragmentTransaction.commit()
         }
         supportActionBar?.setDisplayShowHomeEnabled(true)
         supportActionBar?.setLogo(R.drawable.ic_logo)
         supportActionBar?.setDisplayUseLogoEnabled(true)
+
+        intent?.let {
+            if (it.getBooleanExtra(OPEN_CLIENT_INVESTMENT, false)) {
+                displayClientInvestment(client)
+            }
+        }
     }
 
     override fun displayClientInvestment(client: Client?) {
-        var clientInvestmentFragment = supportFragmentManager.findFragmentByTag(ClientInvestmentFragment.TAG)
+        var clientInvestmentFragment =
+            supportFragmentManager.findFragmentByTag(ClientInvestmentFragment.TAG)
         if (clientInvestmentFragment == null) {
             val fragmentTransaction = supportFragmentManager.beginTransaction()
             clientInvestmentFragment = ClientInvestmentFragment.newInstance(client)
-            fragmentTransaction.add(R.id.agreement_container, clientInvestmentFragment, ClientInvestmentFragment.TAG)
+            fragmentTransaction.add(
+                R.id.agreement_container,
+                clientInvestmentFragment,
+                ClientInvestmentFragment.TAG
+            )
             fragmentTransaction.addToBackStack(ClientInvestmentFragment.TAG)
             fragmentTransaction.commit()
         }
     }
 
     override fun displayAgreementTemplates(client: Client?) {
-        var agreementTemplatesFragment = supportFragmentManager.findFragmentByTag(AgreementTemplatesFragment.TAG)
+        var agreementTemplatesFragment =
+            supportFragmentManager.findFragmentByTag(AgreementTemplatesFragment.TAG)
         if (agreementTemplatesFragment == null) {
             val fragmentTransaction = supportFragmentManager.beginTransaction()
             agreementTemplatesFragment = AgreementTemplatesFragment.newInstance(client)
-            fragmentTransaction.add(R.id.agreement_container, agreementTemplatesFragment, AgreementTemplatesFragment.TAG)
+            fragmentTransaction.add(
+                R.id.agreement_container,
+                agreementTemplatesFragment,
+                AgreementTemplatesFragment.TAG
+            )
             fragmentTransaction.addToBackStack(AgreementTemplatesFragment.TAG)
             fragmentTransaction.commit()
         }