+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/fragment/ClientInvestmentFragment.kt
@@ -8,37 +8,32 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.Spinner
-import android.widget.ArrayAdapter
-import android.widget.CheckBox
-import android.widget.TextView
-import android.widget.AdapterView
-import android.widget.Button
-import android.widget.Toast
-import android.widget.ProgressBar
+import android.widget.*
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import com.docusign.androidsdk.DocuSign
+import com.docusign.androidsdk.dsmodels.DSEnvelope
 import com.docusign.androidsdk.exceptions.DSEnvelopeException
+import com.docusign.androidsdk.listeners.DSCacheEnvelopeListener
 import com.docusign.androidsdk.listeners.DSComposeAndSendEnvelopeListener
+import com.docusign.androidsdk.listeners.DSGetEnvelopeListener
+import com.docusign.esign.api.EnvelopesApi
+import com.docusign.esign.model.EnvelopeSummary
 import com.docusign.sdksamplekotlin.R
 import com.docusign.sdksamplekotlin.SDKSampleApplication
 import com.docusign.sdksamplekotlin.activity.AgreementActivity
-import com.docusign.sdksamplekotlin.livedata.SignOfflineModel
-import com.docusign.sdksamplekotlin.livedata.SignOnlineModel
-import com.docusign.sdksamplekotlin.livedata.Status
+import com.docusign.sdksamplekotlin.livedata.*
 import com.docusign.sdksamplekotlin.model.AccreditedInvestorVerification
 import com.docusign.sdksamplekotlin.model.AccreditedInvestorVerifier
 import com.docusign.sdksamplekotlin.model.Client
-import com.docusign.sdksamplekotlin.utils.Constants
-import com.docusign.sdksamplekotlin.utils.EnvelopeUtils
-import com.docusign.sdksamplekotlin.utils.SigningType
-import com.docusign.sdksamplekotlin.utils.Utils
-import com.docusign.sdksamplekotlin.utils.ClientUtils
+import com.docusign.sdksamplekotlin.utils.*
 import com.docusign.sdksamplekotlin.viewmodel.SigningViewModel
 import com.google.gson.Gson
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 import java.io.File
+
 
 class ClientInvestmentFragment : Fragment() {
 
@@ -62,7 +57,11 @@ class ClientInvestmentFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         return inflater.inflate(R.layout.fragment_client_investment, container, false)
     }
 
@@ -78,95 +77,235 @@ class ClientInvestmentFragment : Fragment() {
                 this, android.R.layout.simple_spinner_item,
                 investments
             )
-            val accreditedInvestorCheckbox = findViewById<CheckBox>(R.id.accredited_investor_checkbox)
+            val accreditedInvestorCheckbox =
+                findViewById<CheckBox>(R.id.accredited_investor_checkbox)
             investmentSpinnerAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
             investmentAmountSpinner.adapter = investmentSpinnerAdapter
             if (client?.storePref == Constants.CLIENT_B_PREF) {
                 if (investments.isNotEmpty()) {
                     investmentAmountSpinner.setSelection(investments.size - 1)
-                    val accreditedInvestorTextView = findViewById<TextView>(R.id.accredited_investor_text_view)
+                    val accreditedInvestorTextView =
+                        findViewById<TextView>(R.id.accredited_investor_text_view)
                     accreditedInvestorTextView.visibility = View.VISIBLE
                     accreditedInvestorCheckbox.visibility = View.VISIBLE
                     accreditedInvestorCheckbox.isChecked = true
                 }
             }
-            investmentAmountSpinner.onItemSelectedListener = object: AdapterView.OnItemSelectedListener{
-                override fun onNothingSelected(p0: AdapterView<*>?) {
-                    /* NO-OP */
-                }
+            investmentAmountSpinner.onItemSelectedListener =
+                object : AdapterView.OnItemSelectedListener {
+                    override fun onNothingSelected(p0: AdapterView<*>?) {
+                        /* NO-OP */
+                    }
 
-                override fun onItemSelected(adapterView: AdapterView<*>?, view: View?, position: Int, id: Long) {
-                    val clientGraphs = arrayOf(R.drawable.growth_portfolio_a_0, R.drawable.growth_portfolio_a_1, R.drawable.growth_portfolio_a_2)
-                    if (position < clientGraphs.size) {
-                        clientGraphImageView.setImageResource(clientGraphs[position])
-                        client?.apply {
-                            investmentAmount = investments[position]
-                            val sharedPreferences = context?.getSharedPreferences(Constants.APP_SHARED_PREFERENCES, Context.MODE_PRIVATE)
-                            val clientJson = Gson().toJson(client)
-                            clientJson?.let {
-                                sharedPreferences?.edit()?.putString(storePref, clientJson)?.apply()
+                    override fun onItemSelected(
+                        adapterView: AdapterView<*>?,
+                        view: View?,
+                        position: Int,
+                        id: Long
+                    ) {
+                        val clientGraphs = arrayOf(
+                            R.drawable.growth_portfolio_a_0,
+                            R.drawable.growth_portfolio_a_1,
+                            R.drawable.growth_portfolio_a_2
+                        )
+                        if (position < clientGraphs.size) {
+                            clientGraphImageView.setImageResource(clientGraphs[position])
+                            client?.apply {
+                                investmentAmount = investments[position]
+                                val sharedPreferences = context?.getSharedPreferences(
+                                    Constants.APP_SHARED_PREFERENCES,
+                                    Context.MODE_PRIVATE
+                                )
+                                val clientJson = Gson().toJson(client)
+                                clientJson?.let {
+                                    sharedPreferences?.edit()?.putString(storePref, clientJson)
+                                        ?.apply()
+                                }
+                                accreditedInvestorCheckbox.isChecked =
+                                    position == investments.size - 1
                             }
-                            accreditedInvestorCheckbox.isChecked = position == investments.size -1
                         }
                     }
                 }
-            }
             val investButton = findViewById<Button>(R.id.invest_button)
             investButton.setOnClickListener {
-                if (client?.storePref == Constants.CLIENT_A_PREF) {
-                    // Launch template flow
-                    (activity as AgreementActivity).displayAgreementTemplates(client)
-                } else {
-                    // Launch envelope flow
-                    createEnvelope(this, accreditedInvestorCheckbox.isChecked, client?.storePref)
+                when (client?.storePref) {
+                    Constants.CLIENT_A_PREF -> {
+                        // Launch template flow
+                        (activity as AgreementActivity).displayAgreementTemplates(client)
+                    }
+                    Constants.CLIENT_B_PREF -> {
+                        // Launch envelope flow
+                        createEnvelope(
+                            this,
+                            accreditedInvestorCheckbox.isChecked,
+                            client?.storePref
+                        )
+                    }
+                    Constants.CLIENT_C_PREF -> {
+                        if (client?.cacheEnvelope == true) {
+                            cachedEnvelope(requireContext(), client?.storePref)
+                        } else {
+                            captiveSigning(this, client?.storePref)
+                        }
+                    }
                 }
             }
         }
     }
 
+    private fun cachedEnvelope(context: Context, clientPref: String?) {
+
+        val envelopeDefinition =
+            EnvelopeUtils.buildCachedEnvelopeDefinition(clientPref!!, requireActivity())
+
+        val envelopesApi =
+            DocuSign.getInstance().getESignApiDelegate().createApiService(EnvelopesApi::class.java)
+
+        val authenticationDelegate = DocuSign.getInstance().getAuthenticationDelegate()
+        val user = authenticationDelegate.getLoggedInUser(context)
+        val call = envelopesApi?.envelopesPostEnvelopes(
+            user.accountId,
+            null,
+            null,
+            null,
+            null,
+            null,
+            envelopeDefinition
+        )
+
+        val envelopeDelegate = DocuSign.getInstance().getEnvelopeDelegate()
+
+        call?.enqueue(object : Callback<EnvelopeSummary> {
+            override fun onResponse(
+                call: Call<EnvelopeSummary>,
+                response: Response<EnvelopeSummary>
+            ) {
+                if (response.isSuccessful) {
+                    envelopeDelegate.cacheEnvelope(
+                        response.body()?.envelopeId ?: "",
+                        object : DSCacheEnvelopeListener {
+                            override fun onComplete(envelope: DSEnvelope) {
+                                signingViewModel.signCachedEnvelope(context, envelope.envelopeId)
+                            }
+
+                            override fun onError(exception: DSEnvelopeException) {
+                                Log.d(OverviewFragment.TAG, exception.message!!)
+                                Toast.makeText(activity, exception.message, Toast.LENGTH_LONG)
+                                    .show()
+                            }
+
+                            override fun onStart() {
+                                toggleProgressBar(true)
+                            }
+                        })
+                }
+            }
+
+            override fun onFailure(call: Call<EnvelopeSummary>, t: Throwable) {
+                Toast.makeText(context, t.message, Toast.LENGTH_LONG).show()
+            }
+        })
+
+    }
+
     private fun initLiveDataObservers(context: Context) {
-        signingViewModel.signOfflineLiveData.observe(viewLifecycleOwner, Observer<SignOfflineModel> { model ->
-            when (model.status) {
-                Status.START -> {
-                    toggleProgressBar(false)
-                }
-                Status.COMPLETE -> {
-                    toggleProgressBar(false)
-                    showSuccessfulSigningDialog(context, SigningType.OFFLINE_SIGNING)
-                    client?.apply {
-                        ClientUtils.setSignedStatus(requireContext(), storePref, true)
+        signingViewModel.signOfflineLiveData.observe(
+            viewLifecycleOwner,
+            Observer<SignOfflineModel> { model ->
+                when (model.status) {
+                    Status.START -> {
+                        toggleProgressBar(false)
+                    }
+                    Status.COMPLETE -> {
+                        toggleProgressBar(false)
+                        showSuccessfulSigningDialog(context, SigningType.OFFLINE_SIGNING)
+                        client?.apply {
+                            ClientUtils.setSignedStatus(requireContext(), storePref, true)
+                        }
+                    }
+                    Status.ERROR -> {
+                        toggleProgressBar(false)
+                        model.exception?.let { exception ->
+                            Log.d(OverviewFragment.TAG, exception.message!!)
+                            Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG)
+                                .show()
+                        }
                     }
                 }
-                Status.ERROR -> {
-                    toggleProgressBar(false)
-                    model.exception?.let { exception ->
-                        Log.d(OverviewFragment.TAG, exception.message!!)
-                        Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG).show()
+            })
+        signingViewModel.signOnlineLiveData.observe(
+            viewLifecycleOwner,
+            Observer<SignOnlineModel> { model ->
+                when (model.status) {
+                    Status.START -> {
+                        toggleProgressBar(false)
+                    }
+                    Status.COMPLETE -> {
+                        toggleProgressBar(false)
+                        showSuccessfulSigningDialog(context, SigningType.ONLINE_SIGNING)
+                        client?.apply {
+                            ClientUtils.setSignedStatus(requireContext(), storePref, true)
+                        }
+                    }
+                    Status.ERROR -> {
+                        toggleProgressBar(false)
+                        model.exception?.let { exception ->
+                            Log.d(OverviewFragment.TAG, exception.message!!)
+                            Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG)
+                                .show()
+                        }
                     }
                 }
-            }
-        })
-        signingViewModel.signOnlineLiveData.observe(viewLifecycleOwner, Observer<SignOnlineModel> { model ->
-            when (model.status) {
-                Status.START -> {
-                    toggleProgressBar(false)
-                }
-                Status.COMPLETE -> {
-                    toggleProgressBar(false)
-                    showSuccessfulSigningDialog(context, SigningType.ONLINE_SIGNING)
-                    client?.apply {
-                        ClientUtils.setSignedStatus(requireContext(), storePref, true)
+            })
+        signingViewModel.captiveSigningLiveData.observe(
+            viewLifecycleOwner,
+            Observer<CaptiveSigningModel> { model ->
+                when (model.status) {
+                    Status.START -> {
+                        toggleProgressBar(false)
+                    }
+                    Status.COMPLETE -> {
+                        toggleProgressBar(false)
+                        showSuccessfulSigningDialog(context, SigningType.CAPTIVE_SIGNING)
+                        client?.apply {
+                            ClientUtils.setSignedStatus(requireContext(), storePref, true)
+                        }
+                    }
+                    Status.ERROR -> {
+                        toggleProgressBar(false)
+                        model.exception?.let { exception ->
+                            Log.d(OverviewFragment.TAG, exception.message!!)
+                            Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG)
+                                .show()
+                        }
                     }
                 }
-                Status.ERROR -> {
-                    toggleProgressBar(false)
-                    model.exception?.let { exception ->
-                        Log.d(OverviewFragment.TAG, exception.message!!)
-                        Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG).show()
+            })
+        signingViewModel.cachedEnvelopeSigningLiveData.observe(
+            viewLifecycleOwner,
+            Observer<CachedEnvelopeSigningModel> { model ->
+                when (model.status) {
+                    Status.START -> {
+                        toggleProgressBar(false)
+                    }
+                    Status.COMPLETE -> {
+                        toggleProgressBar(false)
+                        showSuccessfulSigningDialog(context, SigningType.OFFLINE_SIGNING)
+                        client?.apply {
+                            ClientUtils.setSignedStatus(requireContext(), storePref, true)
+                        }
+                    }
+                    Status.ERROR -> {
+                        toggleProgressBar(false)
+                        model.exception?.let { exception ->
+                            Log.d(OverviewFragment.TAG, exception.message!!)
+                            Toast.makeText(activity, model.exception.message, Toast.LENGTH_LONG)
+                                .show()
+                        }
                     }
                 }
-            }
-        })
+            })
     }
 
     private fun toggleProgressBar(isBusy: Boolean) {
@@ -177,9 +316,63 @@ class ClientInvestmentFragment : Fragment() {
         }
     }
 
-    private fun createEnvelope(context: Context, isAccreditedInvestor: Boolean, clientPref: String?) {
-        val document = if (clientPref == Constants.CLIENT_A_PREF) (activity?.application as SDKSampleApplication).portfolioADoc else
-            (activity?.application as SDKSampleApplication).portfolioBDoc
+    private fun captiveSigning(context: Context, clientPref: String?) {
+        val envelopeDefinition =
+            EnvelopeUtils.buildEnvelopeDefinition(clientPref!!, requireActivity())
+
+        val envelopesApi =
+            DocuSign.getInstance().getESignApiDelegate().createApiService(EnvelopesApi::class.java)
+
+        val authenticationDelegate = DocuSign.getInstance().getAuthenticationDelegate()
+        val user = authenticationDelegate.getLoggedInUser(context)
+        val call = envelopesApi?.envelopesPostEnvelopes(
+            user.accountId,
+            null,
+            null,
+            null,
+            null,
+            null,
+            envelopeDefinition
+        )
+
+        val envelopeDelegate = DocuSign.getInstance().getEnvelopeDelegate()
+
+        call?.enqueue(object : Callback<EnvelopeSummary> {
+            override fun onResponse(
+                call: Call<EnvelopeSummary>,
+                response: Response<EnvelopeSummary>
+            ) {
+                if (response.isSuccessful) {
+                    val envelopeSummary = response.body()
+                    envelopeSummary?.let {
+                        envelopeDelegate.getEnvelope(it.envelopeId, object : DSGetEnvelopeListener {
+                            override fun onError(exception: DSEnvelopeException) {
+                                Toast.makeText(context, exception.message, Toast.LENGTH_LONG).show()
+                            }
+
+                            override fun onSuccess(envelope: DSEnvelope) {
+                                signingViewModel.captiveSigning(requireContext(), envelope)
+                            }
+                        })
+                    }
+
+                }
+            }
+
+            override fun onFailure(call: Call<EnvelopeSummary>, t: Throwable) {
+                Toast.makeText(context, t.message, Toast.LENGTH_LONG).show()
+            }
+        })
+    }
+
+    private fun createEnvelope(
+        context: Context,
+        isAccreditedInvestor: Boolean,
+        clientPref: String?
+    ) {
+        val document =
+            if (clientPref == Constants.CLIENT_A_PREF) (activity?.application as SDKSampleApplication).portfolioADoc else
+                (activity?.application as SDKSampleApplication).portfolioBDoc
 
         if (document == null) {
             Log.e(OverviewFragment.TAG, "Unable to retrieve document")
@@ -190,13 +383,19 @@ class ClientInvestmentFragment : Fragment() {
 
         if (isAccreditedInvestor) {
             var accreditedInvestorVerificationDocument: File? = null
-            accreditedInvestorVerificationDocument = (activity?.application as SDKSampleApplication).accreditedInvestorDoc
+            accreditedInvestorVerificationDocument =
+                (activity?.application as SDKSampleApplication).accreditedInvestorDoc
 
             if (accreditedInvestorVerificationDocument == null) {
-                Toast.makeText(context, "AccreditedInvestor Verification document is not available", Toast.LENGTH_LONG).show()
+                Toast.makeText(
+                    context,
+                    "AccreditedInvestor Verification document is not available",
+                    Toast.LENGTH_LONG
+                ).show()
                 return
             }
-            val sharedPreferences = context.getSharedPreferences(Constants.APP_SHARED_PREFERENCES, Context.MODE_PRIVATE)
+            val sharedPreferences =
+                context.getSharedPreferences(Constants.APP_SHARED_PREFERENCES, Context.MODE_PRIVATE)
             val clientString = sharedPreferences.getString(clientPref, null)
 
             if (clientString == null) {
@@ -205,44 +404,59 @@ class ClientInvestmentFragment : Fragment() {
             }
 
             val client = Gson().fromJson<Client>(clientString, Client::class.java)
-            val clientAddress = client.addressLine1 + "," + client.addressLine2 + "," + client.addressLine3
+            val clientAddress =
+                client.addressLine1 + "," + client.addressLine2 + "," + client.addressLine3
             val authenticationDelegate = DocuSign.getInstance().getAuthenticationDelegate()
             val user = authenticationDelegate.getLoggedInUser(context)
-            val accreditedInvestorVerifier = AccreditedInvestorVerifier(user.name,
+            val accreditedInvestorVerifier = AccreditedInvestorVerifier(
+                user.name,
                 getString(R.string.tgkfinancial), "TN12345", "CA",
-                "202 Main Street", null, "San Francisco, CA, 94105")
+                "202 Main Street", null, "San Francisco, CA, 94105"
+            )
             accreditedInvestorVerification =
-                AccreditedInvestorVerification(client.name, clientAddress, accreditedInvestorVerifier, accreditedInvestorVerificationDocument)
+                AccreditedInvestorVerification(
+                    client.name,
+                    clientAddress,
+                    accreditedInvestorVerifier,
+                    accreditedInvestorVerificationDocument
+                )
 
         }
-        val envelope = EnvelopeUtils.buildEnvelope(context, document, accreditedInvestorVerification, clientPref)
+        val envelope = EnvelopeUtils.buildEnvelope(
+            context,
+            document,
+            accreditedInvestorVerification,
+            clientPref
+        )
         if (envelope == null) {
             Log.e(OverviewFragment.TAG, "Unable to create envelope")
             Toast.makeText(context, "Unable to create envelope", Toast.LENGTH_LONG).show()
             return
         }
         val envelopeDelegate = DocuSign.getInstance().getEnvelopeDelegate()
-        envelopeDelegate.composeAndSendEnvelope(envelope, object : DSComposeAndSendEnvelopeListener {
+        envelopeDelegate.composeAndSendEnvelope(
+            envelope,
+            object : DSComposeAndSendEnvelopeListener {
 
-            override fun onSuccess(envelopeId: String, isEnvelopeSent: Boolean) {
-                if (Utils.isNetworkAvailable(context)) {
-                    toggleProgressBar(true)
-                    signingViewModel.signOnline(context, envelopeId)
-                } else {
-                    signingViewModel.signOffline(context, envelopeId)
+                override fun onSuccess(envelopeId: String, isEnvelopeSent: Boolean) {
+                    if (Utils.isNetworkAvailable(context)) {
+                        toggleProgressBar(true)
+                        signingViewModel.signOnline(context, envelopeId)
+                    } else {
+                        signingViewModel.signOffline(context, envelopeId)
+                    }
                 }
-            }
 
-            override fun onError(exception: DSEnvelopeException) {
-                Log.e(OverviewFragment.TAG, exception.message!!)
-                Toast.makeText(context, exception.message, Toast.LENGTH_LONG).show()
-            }
-        })
+                override fun onError(exception: DSEnvelopeException) {
+                    Log.e(OverviewFragment.TAG, exception.message!!)
+                    Toast.makeText(context, exception.message, Toast.LENGTH_LONG).show()
+                }
+            })
     }
 
     private fun showSuccessfulSigningDialog(context: Context, signingType: SigningType) {
         var message = getString(R.string.envelope_signed_offline_message)
-        if (signingType == SigningType.ONLINE_SIGNING) {
+        if (signingType == SigningType.ONLINE_SIGNING || signingType == SigningType.CAPTIVE_SIGNING) {
             message = getString(R.string.envelope_signed_message)
         }
         AlertDialog.Builder(context)