+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/livedata/SigningLiveData.kt
@@ -4,3 +4,5 @@ import com.docusign.androidsdk.exceptions.DSSigningException
 
 data class SignOfflineModel(val status: Status, val exception: DSSigningException?)
 data class SignOnlineModel(val status: Status, val exception: DSSigningException?)
+data class CaptiveSigningModel(val status: Status, val exception: DSSigningException?)
+data class CachedEnvelopeSigningModel(val status: Status, val exception: DSSigningException?)