+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/utils/ClientUtils.kt
@@ -7,10 +7,16 @@ object ClientUtils {
     fun setSignedStatus(context: Context, clientPref: String, signed: Boolean = false) {
         val sharedPreferences = context.getSharedPreferences(Constants.APP_SHARED_PREFERENCES,
                 Context.MODE_PRIVATE)
-        val signedStatus = if (clientPref == Constants.CLIENT_A_PREF) {
-            Constants.CLIENT_A_SIGNED_STATUS
-        } else {
-            Constants.CLIENT_B_SIGNED_STATUS
+        val signedStatus = when (clientPref) {
+            Constants.CLIENT_A_PREF -> {
+                Constants.CLIENT_A_SIGNED_STATUS
+            }
+            Constants.CLIENT_B_PREF -> {
+                Constants.CLIENT_B_SIGNED_STATUS
+            }
+            else -> {
+                Constants.CLIENT_C_SIGNED_STATUS
+            }
         }
         sharedPreferences.edit().putBoolean(signedStatus, signed).apply()
     }