+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/utils/EnvelopeUtils.kt
@@ -1,6 +1,8 @@
 package com.docusign.sdksamplekotlin.utils
 
+import android.app.Activity
 import android.content.Context
+import android.util.Base64
 import android.util.Log
 import android.widget.Toast
 import com.docusign.androidsdk.DocuSign
@@ -19,6 +21,8 @@ import com.docusign.androidsdk.dsmodels.DSEnvelopeDefaultsUniqueRecipientSelecto
 import com.docusign.androidsdk.exceptions.DSAuthenticationException
 import com.docusign.androidsdk.exceptions.DSEnvelopeException
 import com.docusign.androidsdk.exceptions.DocuSignNotInitializedException
+import com.docusign.esign.model.*
+import com.docusign.sdksamplekotlin.SDKSampleApplication
 import com.docusign.sdksamplekotlin.model.AccreditedInvestorVerification
 import com.docusign.sdksamplekotlin.model.Client
 import com.google.gson.Gson
@@ -29,6 +33,127 @@ import java.util.ArrayList
 object EnvelopeUtils {
 
     private val TAG = EnvelopeUtils::class.java.simpleName
+
+    fun buildEnvelopeDefinition(clientPref: String, activity: Activity) : EnvelopeDefinition? {
+        val document =
+            if (clientPref == Constants.CLIENT_A_PREF) (activity.application as SDKSampleApplication).portfolioADoc else
+                (activity.application as SDKSampleApplication).portfolioBDoc
+
+        val buffer: ByteArray = document?.readBytes()!!
+        val envelopeDefinition = EnvelopeDefinition()
+        envelopeDefinition.emailSubject = "Subject"
+        val doc1 = Document()
+        val doc1b64: String = Base64.encodeToString(buffer, Base64.DEFAULT)
+        doc1.documentBase64 = doc1b64
+        doc1.name = "Lorem Ipsum"
+        doc1.fileExtension = "pdf"
+        doc1.documentId = "3"
+        envelopeDefinition.documents = listOf(doc1)
+
+        val sharedPreferences =
+            activity.getSharedPreferences(Constants.APP_SHARED_PREFERENCES, Context.MODE_PRIVATE)
+        val clientString = sharedPreferences.getString(clientPref, null)
+
+        if (clientString == null) {
+            Toast.makeText(activity, "Client details not available", Toast.LENGTH_LONG).show()
+            return null
+        }
+
+        val client = Gson().fromJson<Client>(clientString, Client::class.java)
+        val signer1 = InPersonSigner()
+        signer1.email = client.email
+        signer1.name = client.name
+        signer1.signerName = client.name
+        signer1.signerEmail = client.email
+        signer1.clientUserId(client.id)
+        signer1.recipientId("1")
+
+        val signHere1 = SignHere()
+        signHere1.documentId = "3"
+        signHere1.recipientId = "1"
+        signHere1.pageNumber = "1"
+        signHere1.xPosition = "370"
+        signHere1.yPosition = "720"
+
+        val signer1Tabs = Tabs()
+        signer1Tabs.textTabs = getEnvelopeDefinitionTextTabs(client)
+        signer1Tabs.signHereTabs = listOf(signHere1)
+        signer1.tabs = signer1Tabs
+
+        val recipients = Recipients()
+
+        recipients.addInPersonSignersItem(signer1)
+        envelopeDefinition.recipients = recipients
+
+        envelopeDefinition.status = "sent"
+
+        return envelopeDefinition
+    }
+
+    fun buildCachedEnvelopeDefinition(clientPref: String, activity: Activity): EnvelopeDefinition? {
+        val document =
+            if (clientPref == Constants.CLIENT_A_PREF) (activity.application as SDKSampleApplication).portfolioADoc else
+                (activity.application as SDKSampleApplication).portfolioBDoc
+
+        val buffer: ByteArray = document?.readBytes()!!
+        val envelopeDefinition = EnvelopeDefinition()
+        envelopeDefinition.emailSubject = "Subject"
+        val doc1 = Document()
+        val doc1b64: String = Base64.encodeToString(buffer, Base64.DEFAULT)
+        doc1.documentBase64 = doc1b64
+        doc1.name = "Lorem Ipsum"
+        doc1.fileExtension = "pdf"
+        doc1.documentId = "3"
+        doc1.order = "1"
+        envelopeDefinition.documents = listOf(doc1)
+
+        val sharedPreferences =
+            activity.getSharedPreferences(Constants.APP_SHARED_PREFERENCES, Context.MODE_PRIVATE)
+        val clientString = sharedPreferences.getString(clientPref, null)
+
+        if (clientString == null) {
+            Toast.makeText(activity, "Client details not available", Toast.LENGTH_LONG).show()
+            return null
+        }
+
+        val client = Gson().fromJson<Client>(clientString, Client::class.java)
+
+        val user = DocuSign.getInstance().getAuthenticationDelegate().getLoggedInUser(activity)
+
+        val signer1 = InPersonSigner()
+        signer1.recipientId("1")
+        signer1.signingGroupId = null
+        signer1.signingGroupName = null
+        signer1.canSignOffline = "true"
+        signer1.routingOrder = "1"
+        signer1.recipientType = DSRecipientType.IN_PERSON_SIGNER.name
+        signer1.hostEmail = user.email
+        signer1.hostName = user.name
+        signer1.signerName = client.name
+        signer1.recipientId = "1"
+
+        val signHere1 = SignHere()
+        signHere1.documentId = "3"
+        signHere1.recipientId = "1"
+        signHere1.pageNumber = "1"
+        signHere1.xPosition = "370"
+        signHere1.yPosition = "720"
+
+        val signer1Tabs = Tabs()
+        signer1Tabs.textTabs = getEnvelopeDefinitionTextTabs(client)
+        signer1Tabs.signHereTabs = listOf(signHere1)
+        signer1.tabs = signer1Tabs
+
+        val recipients = Recipients()
+        recipients.inPersonSigners = listOf(signer1)
+        envelopeDefinition.recipients = recipients
+
+        envelopeDefinition.is21CFRPart11 = "false"
+        envelopeDefinition.signerCanSignOnMobile = "true"
+        envelopeDefinition.status = "sent"
+        envelopeDefinition.emailSubject = "Test Envelope Offline Signing"
+        return envelopeDefinition
+    }
 
     fun buildEnvelope(
         context: Context,
@@ -280,7 +405,7 @@ object EnvelopeUtils {
         return tabs
     }
 
-    private fun createInvestmentAgreementTabs(client: Client): List<DSTab> {
+    fun createInvestmentAgreementTabs(client: Client): List<DSTab> {
         val tabs = mutableListOf<DSTab>()
         tabs.add(
             DSTab.Builder()
@@ -376,6 +501,82 @@ object EnvelopeUtils {
                 .optional(true)
                 .build()
         )                        // Name of the applicant
+        return tabs
+    }
+
+    private fun getEnvelopeDefinitionTextTabs(client: Client) : List<Text> {
+        val tabs = mutableListOf<Text>()
+
+        val addressLine1 = Text()
+        addressLine1.documentId = "3"
+        addressLine1.recipientId = "1"
+        addressLine1.pageNumber = "1"
+        addressLine1.xPosition = "370"
+        addressLine1.yPosition = "110"
+        addressLine1.value = client.addressLine1
+
+        tabs.add(addressLine1)
+
+        val addressLine2 = Text()
+        addressLine2.documentId = "3"
+        addressLine2.recipientId = "1"
+        addressLine2.pageNumber = "1"
+        addressLine2.xPosition = "370"
+        addressLine2.yPosition = "140"
+        addressLine2.value = client.addressLine2
+
+        tabs.add(addressLine2)
+
+        val addressLine3 = Text()
+        addressLine3.documentId = "3"
+        addressLine3.recipientId = "1"
+        addressLine3.pageNumber = "1"
+        addressLine3.xPosition = "370"
+        addressLine3.yPosition = "165"
+        addressLine3.value = client.addressLine3
+
+        tabs.add(addressLine3)
+
+        val name = Text()
+        name.documentId = "3"
+        name.recipientId = "1"
+        name.pageNumber = "1"
+        name.xPosition = "50"
+        name.yPosition = "165"
+        name.value = client.name
+
+        tabs.add(name)
+
+        val clientId = Text()
+        clientId.documentId = "3"
+        clientId.recipientId = "1"
+        clientId.pageNumber = "1"
+        clientId.xPosition = "370"
+        clientId.yPosition = "550"
+        clientId.value = client.id
+
+        tabs.add(clientId)
+
+        val investmentAmount = Text()
+        investmentAmount.documentId = "3"
+        investmentAmount.recipientId = "1"
+        investmentAmount.pageNumber = "1"
+        investmentAmount.xPosition = "370"
+        investmentAmount.yPosition = "630"
+        investmentAmount.value = client.investmentAmount
+
+        tabs.add(investmentAmount)
+
+        val signatureName = Text()
+        signatureName.documentId = "3"
+        signatureName.recipientId = "1"
+        signatureName.pageNumber = "1"
+        signatureName.xPosition = "50"
+        signatureName.yPosition = "730"
+        signatureName.value = client.name
+
+        tabs.add(signatureName)
+
         return tabs
     }