+++ docusign-sdk-sample-kotlin/app/src/main/java/com/docusign/sdksamplekotlin/utils/SigningType.kt
@@ -2,5 +2,6 @@ package com.docusign.sdksamplekotlin.utils
 
 enum class SigningType {
     ONLINE_SIGNING,
-    OFFLINE_SIGNING
+    OFFLINE_SIGNING,
+    CAPTIVE_SIGNING
 }